feat: add basic config parser
A basic config parser has been implemented in `config.rs`. It is still
incomplete, but one letter key press bindings in QWERTYUIOP should work
now, along with a command on the next line.

Example `swhkdrc` file that should work:

```
r
    alacritty

w
    kitty

t
    /bin/firefox
```

**TODO:**

- Add more supported key presses (alphabetical, numerical, insert,
        home, delete, end, etc)
- Add chained keybindings (`super + g`, `super + 3`)
- Add error handling for parsing the contents of the file

Note: `server.rs` has been modified to add `mod config` at the start of
the file, but it was necessary to do testing for the config parser.