feat(daemon): swallow key event for hotkeys
swhkd now swallows the key events if they're used by the hotkeys,
for example: if you define a hotkey as `super + r` in the config and
press `super + r`, your WM/compositor will not register the r key.

**WARNING:** have a keyboard that you can plug/unplug in this stage
of development, because if you exit swhkd, the keyboard device will
still remain grabbed and you will have to plug / unplug them to reset them.