[fix] raise swhkd privileges right after reading config
As a fix for CVE-2022-27814, root privileges are dropped to the calling user
when (re)loading the config file. Privileges were sometimes dropped but never
raised again, which caused crashes when sending SIGHUP to swhkd multiple times
in a row.

This now always raises privileges after successfully reading the config file.
Fixes #155.