[feat] add polkit policy file generation script
The build-polkit-policy.sh script (in the newly added 'scripts' folder) builds a
polkit policy file for swhkd. It has a number of optional arguments to make it
easy to extend and adapt to different use cases. The output path, the swhkd
binary path, and even the custom authentication message and the action id are
configurable.