fix: unblock all rfkill softblocks to avoid radio devs turning off
This commit attempts at fixing the major issue of rfkill not being
properly turned off while creating the swhkd virtual device.

Changes made:

1. Added RfkillEvent struct that stores the rfkill details and status
required with documentation
2. Added two new macros with `ioctl_readwrite` to ensure that we can
query the rfkill status and modify properly
3. New function `unblock_rfkill` that unblocks all soft blocks for any
device
4. Called `unblock_rfkill` function in `create_uinput_switches_device`
that ensures that rfkill is NOT blocked and complies with rfkill-input
being disabled

This should ensure that rfkill does not react when virtual device is
being created.