fix: refined swhks ipc.rs for hash calculation and finding buffer byte (#306)
* fix: unblock all rfkill softblocks to avoid radio devs turning off

This commit attempts at fixing the major issue of rfkill not being
properly turned off while creating the swhkd virtual device.

Changes made:

1. Added RfkillEvent struct that stores the rfkill details and status
required with documentation
2. Added two new macros with `ioctl_readwrite` to ensure that we can
query the rfkill status and modify properly
3. New function `unblock_rfkill` that unblocks all soft blocks for any
device
4. Called `unblock_rfkill` function in `create_uinput_switches_device`
that ensures that rfkill is NOT blocked and complies with rfkill-input
being disabled

This should ensure that rfkill does not react when virtual device is
being created.

* fix: clean up code in swhks ipc.rs

1. Calculate hash function now borrows the string instead of taking
ownership
2. Used match to check for buffer byte instead of if statement
3. Added more debug logging and error handling

* chore: fallback commit to parent ca82393

Remove updated logic for uinput.rs as it does not work

* fix: use is_some_and instead of map_for in daemon.rs

replaced `map_for` with `is_some_and` as it is more efficient in this
case (simple bool ops)